--- a/cpp/arcticdb/codec/segment.hpp
+++ b/cpp/arcticdb/codec/segment.hpp
@@ -7,12 +7,9 @@
 
 #pragma once
 
-#include "util/buffer.hpp"
+#include <arcticdb/util/buffer.hpp>
 #include <arcticdb/storage/common.hpp>
-#include <codec/encoding_sizes.hpp>
-#include <google/protobuf/io/zero_copy_stream_impl.h>
 #include <google/protobuf/arena.h>
-#include <util/buffer_pool.hpp>
 #include <arcticdb/entity/field_collection.hpp>
 
 #include <iostream>