--- a/cpp/arcticdb/pipeline/column_stats.hpp
+++ b/cpp/arcticdb/pipeline/column_stats.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <arcticdb/processing/clause.hpp>
-#include <arcticdb/entity/protobufs.hpp>
 #include <ankerl/unordered_dense.h>
 
 #include <map>