--- a/cpp/arcticdb/pipeline/python_output_frame.hpp
+++ b/cpp/arcticdb/pipeline/python_output_frame.hpp
@@ -8,7 +8,6 @@
 #pragma once
 
 #include <arcticdb/pipeline/frame_data_wrapper.hpp>
-#include <arcticdb/util/buffer_holder.hpp>
 
 namespace arcticdb::pipelines {