… .github/workflows/analysis_workflow.yml
… .github/workflows/build.yml
… cpp/arcticdb/CMakeLists.txt
… cpp/arcticdb/async/test/test_async.cpp
… cpp/arcticdb/codec/test/test_codec.cpp
… cpp/arcticdb/column_store/test/benchmark_memory_segment.cpp
… cpp/arcticdb/column_store/test/ingestion_stress_test.cpp
… cpp/arcticdb/column_store/test/rapidcheck_chunked_buffer.cpp
… cpp/arcticdb/column_store/test/rapidcheck_column.cpp
… cpp/arcticdb/column_store/test/rapidcheck_column_data_random_accessor.cpp
… cpp/arcticdb/column_store/test/rapidcheck_column_map.cpp
… cpp/arcticdb/column_store/test/test_column.cpp
… cpp/arcticdb/column_store/test/test_index_filtering.cpp
… cpp/arcticdb/pipeline/test/test_pipeline.cpp
… cpp/arcticdb/processing/clause.cpp
… cpp/arcticdb/processing/clause.hpp
… cpp/arcticdb/processing/component_manager.hpp
… cpp/arcticdb/processing/operation_dispatch.cpp
… cpp/arcticdb/processing/operation_dispatch.hpp
… cpp/arcticdb/processing/operation_dispatch_binary.cpp
… cpp/arcticdb/processing/operation_dispatch_binary.hpp
… cpp/arcticdb/processing/operation_dispatch_binary_eq.cpp
… cpp/arcticdb/processing/operation_dispatch_binary_gt.cpp
… cpp/arcticdb/processing/operation_dispatch_binary_lt.cpp
… cpp/arcticdb/processing/operation_dispatch_binary_operator.cpp
… cpp/arcticdb/processing/operation_dispatch_unary.hpp
… cpp/arcticdb/processing/processing_unit.cpp
… cpp/arcticdb/processing/processing_unit.hpp
… cpp/arcticdb/processing/test/benchmark_clause.cpp
… cpp/arcticdb/processing/test/test_clause.cpp