- Fix compilation errors with clang-18 due to Werror
- Run include-what-you-use to reduce includes and add forward declarations

Same as: #1474 in the original PR ASV fails to perform some git commands, the issue seems to be related to branch naming.

**TLDR**
Using the suggestions from [include-what-you-use](https://github.com/include-what-you-use/include-what-you-use) makes full builds ~10% faster sparing 20s on a local machine (Processor: 13th Gen Intel(R) Core(TM) i7-1370P, 1900 Mhz, 14 Core(s), 20 Logical Processor(s)) building with 20 threads.

Build info:
OS: Ubuntu (WSL)
Compiler: Clang 18
Linker: Mold
Build system: Ninja
(Clang and Mold are custom builds using additional flags -O3 -march=native -flto=full, arcticdb is built as usual)
| #      | Master (debug) in seconds  | IWYU (debug) in seconds  |
|--------|-------------|-------------|
| 1      |      212.31 |      203.37 |
| 2      |      234.02 |      212.33 |
| 3      |      234.24 |      213.05 |
| 4      |      233.63 |      213.32 |
| 5      |      233.27 |      213.76 |
| 6      |      233.79 |      213.07 |
| 7      |      234.13 |      213.23 |
| 8      |      232.67 |      212.27 |
| 9      |      233.18 |      212.25 |
| 10     |      233.87 |      212.58 |
| 11     |      233.75 |      213.21 |
| 12     |      233.63 |      213.16 |
| 13     |      232.89 |      213.87 |
| 14     |      233.64 |      213.39 |
| 15     |      233.12 |      213.54 |
| Mean   | 232.1426667 | 212.4266667 |
| Median |      233.63 |      213.16 |
| SD     | 5.505148845 | 2.557028595 |

#### Reference Issues/PRs
<!--Example: Fixes #1234. See also #3456.-->

#### What does this implement or fix?

#### Any other comments?

#### Checklist

<details>
  <summary>
   Checklist for code changes...
  </summary>
 
 - [ ] Have you updated the relevant docstrings, documentation and copyright notice?
 - [ ] Is this contribution tested against [all ArcticDB's features](../docs/mkdocs/docs/technical/contributing.md)?
 - [ ] Do all exceptions introduced raise appropriate [error messages](https://docs.arcticdb.io/error_messages/)?
 - [ ] Are API changes highlighted in the PR description?
 - [ ] Is the PR labelled as enhancement or bug so it appears in autogenerated release notes?
</details>

<!--
Thanks for contributing a Pull Request to ArcticDB! Please ensure you have taken a look at:
 - ArcticDB's Code of Conduct: https://github.com/man-group/ArcticDB/blob/master/CODE_OF_CONDUCT.md
 - ArcticDB's Contribution Licensing: https://github.com/man-group/ArcticDB/blob/master/docs/mkdocs/docs/technical/contributing.md#contribution-licensing
-->