@vasil-pashov 2024-04-27
Checks are failing due to ASV failing to perform some git commands. The bug is reported [here](https://github.com/airspeed-velocity/asv/issues/1399)